+++ src/main/java/life/wpx1997/community/dto/MessageTagDTO.java
@@ -3,10 +3,11 @@ package life.wpx1997.community.dto;
 import lombok.Data;
 
 /**
- * created by 小case on 2019/9/7 23:43
+ * created on 2019/9/7 23:43
+ * @author 不会飞的小鹏
  */
 @Data
 public class MessageTagDTO {
     private String name;
-    private Integer priority;
+    private Long priority;
 }
